completed import and init
