+++ src/e2e-helper/e2e-bitmap-helper.ts
@@ -61,6 +61,26 @@ export default class BitMapHelper {
     fs.ensureFileSync(bitmapFile);
     return fs.writeJsonSync(bitmapFile, bitmap, { spaces: 2 });
   }
+  createHarmony(
+    cwd: string = this.scopes.localPath,
+    componentObject = {
+      'bar/foo': {
+        scope: '',
+        version: '',
+        mainFile: 'bar/foo.js',
+        rootDir: 'bar',
+      },
+    }
+  ) {
+    const bitmapFile = path.join(cwd, BIT_MAP);
+
+    const bitmap = {
+      version: '0.11.1-testing',
+    };
+    Object.keys(componentObject).forEach((key) => (bitmap[key] = componentObject[key]));
+    fs.ensureFileSync(bitmapFile);
+    return fs.writeJsonSync(bitmapFile, bitmap, { spaces: 2 });
+  }
   printFilesInCaseOfError(files: Record<string, any>[]): string {
     // @ts-ignore AUTO-ADDED-AFTER-MIGRATION-PLEASE-FIX!
     const filesStr = files.map((f) => f.name).join(', ');
